requirements.txt
absl-py==0.10.0
aiohttp==3.6.2
alembic==1.4.2
alpaca-trade-api==0.50.1
appnope==0.1.0
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
backcall==0.2.0
bcolz==0.12.1
Bottleneck==1.3.2
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
contextlib2==0.6.0.post1
cyordereddict==1.0.0
Cython==0.29.21
decorator==4.4.2
empyrical==0.5.3
Flask==1.1.2
Flask-Cors==3.0.9
gast==0.3.3
google-auth==1.22.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
idna-ssl==1.1.0
iexfinance==0.4.3
importlib-metadata==2.0.0
intervaltree==3.1.0
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
Keras-Preprocessing==1.1.2
Logbook==1.5.3
lru-dict==1.1.6
lxml==4.5.2
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
multidict==4.7.6
multipledispatch==0.6.0
networkx==1.11
numexpr==2.7.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.44
opt-einsum==3.3.0
pandas==0.22.0
pandas-datareader==0.9.0
parso==0.7.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
pipeline-live==0.1.9
prompt-toolkit==1.0.14
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
PyInquirer==1.0.3
pylivetrader==0.4.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
regex==2020.9.27
requests==2.24.0
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.2
six==1.15.0
sortedcontainers==2.2.2
SQLAlchemy==1.3.19
statsmodels==0.12.0
tables==3.6.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-cpu==2.3.1
tensorflow-estimator==2.3.0
tensorflow-hub==0.7.0
tensorflowjs==2.4.0
termcolor==1.1.0
toolz==0.10.0
trading-calendars==1.11.11
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.0
zipline==1.3.0
zipp==3.2.0
